- Migrated a lot of functions to avoid inline hooking instead of DynamicApiInvoke   - more syscalls, less dynamic invocation, more module mapping (fresh loading) - changed m2 injection to createthread on suspended process for a more hidden and constistency on the iexplore - changed m3 injection to use edge because svchost traffic and c2 traffic is sussy - can now specify a url on /f: to grab a file from a url   - modified how it determines if /f is valid in a better way cuz it was dookie before xd   - fixed file check when loading from a local file - Moved delegates to its own file to clean things up   - also made DynamicInvoke.native.delegates struct public so i can access the nt delegates without rewriting them (but i already rewrote some xd) - made a structs file to hold structures for less overhead
