+++ DinvokeWinCallback/DynamicInvokeGeneric.cs
@@ -1,4 +1,4 @@
-// Author: Ryan Cobb (@cobbr_io)
+﻿// Author: Ryan Cobb (@cobbr_io)
 // Project: SharpSploit (https://github.com/cobbr/SharpSploit)
 // License: BSD 3-Clause
