+++ DinvokeWinCallback/ManualMapMap.cs
