+++ DinvokeWinCallback/Properties/AssemblyInfo.cs
